added `and` operator

+++ parser.cpp
@@ -45,7 +45,7 @@ void Tokenizor::set_stream(FILE *_stream) {
 
 void str_to_lower(string &str) {
     size_t len = str.length();
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
         if ('A' <= str[i] && str[i] <= 'Z')
             str[i] -= 'A' - 'a';
 }
